format functions moved to AuxMethods
