+++ src/additional/AuxiliaryMethods.hpp
@@ -32,5 +32,32 @@ std::string removeSpaces(const std::string& i_s);
 int         skipSpaces(const std::string& i_s, int i_start = 0);
 // TODO: if need CopyDirectory
 std::string intToStringWithZeroes(int i_num, int i_totalDigits = 5);
-// namespace end
+
+template<class Tuple, std::size_t N>
+struct TuplePrinter {
+  static void print(const std::string& fmt, std::ostream& os, const Tuple& t) {
+    const size_t idx = fmt.find_last_of('%');
+    TuplePrinter<Tuple, N - 1>::print(std::string(fmt, 0, idx), os, t);
+    os << std::get<N - 1>(t) << std::string(fmt, idx + 1);
+  }
+};
+
+template<class Tuple>
+struct TuplePrinter<Tuple, 1> {
+  static void print(const std::string& fmt, std::ostream& os, const Tuple& t) {
+    const size_t idx = fmt.find_first_of('%');
+    os << std::string(fmt, 0, idx) << std::get<0>(t)
+       << std::string(fmt, idx + 1);
+  }
+};
+
+template<class... Args>
+std::string format(const std::string& fmt, Args&&... args) {
+  std::stringstream ss;
+
+  const auto        t = std::make_tuple(std::forward<Args>(args)...);
+
+  TuplePrinter<decltype(t), sizeof...(Args)>::print(fmt, ss, t);
+  return ss.str();
+}
 }  // namespace AuxMethods
