+++ src/baseStructures/graph/OrientedGraph.hpp
@@ -161,31 +161,3 @@ private:
 
   std::shared_ptr<Settings> d_settings = Settings::getInstance("OrientedGraph");
 };
-
-template<class Tuple, std::size_t N>
-struct TuplePrinter {
-  static void print(const std::string& fmt, std::ostream& os, const Tuple& t) {
-    const size_t idx = fmt.find_last_of('%');
-    TuplePrinter<Tuple, N - 1>::print(std::string(fmt, 0, idx), os, t);
-    os << std::get<N - 1>(t) << std::string(fmt, idx + 1);
-  }
-};
-
-template<class Tuple>
-struct TuplePrinter<Tuple, 1> {
-  static void print(const std::string& fmt, std::ostream& os, const Tuple& t) {
-    const size_t idx = fmt.find_first_of('%');
-    os << std::string(fmt, 0, idx) << std::get<0>(t)
-       << std::string(fmt, idx + 1);
-  }
-};
-
-template<class... Args>
-std::string format(const std::string& fmt, Args&&... args) {
-  std::stringstream ss;
-
-  const auto        t = std::make_tuple(std::forward<Args>(args)...);
-
-  TuplePrinter<decltype(t), sizeof...(Args)>::print(fmt, ss, t);
-  return ss.str();
-}
